Refactor ShipperTransportationId
https://github.com/metasfresh/metasfresh/issues/5885
