+++ de.metas.shipper.gateway.derkurier/src/main/java/de/metas/shipper/gateway/derkurier/misc/DerKurierDeliveryOrderEmailer.java
@@ -16,7 +16,7 @@ import de.metas.email.MailService;
 import de.metas.email.mailboxes.Mailbox;
 import de.metas.i18n.IMsgBL;
 import de.metas.shipper.gateway.derkurier.DerKurierConstants;
-import de.metas.shipping.api.ShipperTransportationId;
+import de.metas.shipping.model.ShipperTransportationId;
 import de.metas.shipping.model.I_M_ShipperTransportation;
 import de.metas.util.Check;
 import de.metas.util.Services;
