+++ de.metas.shipper.gateway.derkurier/src/test/java/de/metas/shipper/gateway/derkurier/DerKurierTestTools.java
@@ -17,7 +17,7 @@ import de.metas.shipper.gateway.spi.model.DeliveryPosition;
 import de.metas.shipper.gateway.spi.model.OrderId;
 import de.metas.shipper.gateway.spi.model.PickupDate;
 import de.metas.shipping.ShipperId;
-import de.metas.shipping.api.ShipperTransportationId;
+import de.metas.shipping.model.ShipperTransportationId;
 
 import java.time.LocalDate;
 import java.time.LocalTime;
