+++ de.metas.swat/de.metas.swat.base/src/main/java/de/metas/shipping/api/impl/ShipperTransportationDAO.java
@@ -33,7 +33,7 @@ import org.compiere.model.I_M_Package;
 
 import de.metas.shipping.ShipperId;
 import de.metas.shipping.api.IShipperTransportationDAO;
-import de.metas.shipping.api.ShipperTransportationId;
+import de.metas.shipping.model.ShipperTransportationId;
 import de.metas.shipping.model.I_M_ShipperTransportation;
 import de.metas.shipping.model.I_M_ShippingPackage;
 import de.metas.shipping.model.X_M_ShipperTransportation;
